Rename Intra to fork
